# docs/swagger.yaml
definitions:
...
  pagination.Response-user_UserResponseDto:
    properties:
      items:
        items:
          $ref: '#/definitions/user.UserResponseDto'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  rest_err.RestErr:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
      ray_trace:
        type: string
    type: object
  user.AtribuicaoResponseDto:
    properties:
      created_at:
        type: string
      papel:
        type: string
      papel_uuid:
        type: string
      user_uuid:
        type: string
      uuid:
        type: string
      workspace_uuid:
        type: string
    type: object
  user.AtribuirPapelRequestDto:
    properties:
      papel_uuid:
        type: string
      workspace_uuid:
        type: string
    required:
    - papel_uuid
    - workspace_uuid
    type: object
  user.CreateUserRequestDto:
    properties:
      email:
        maxLength: 254
        type: string
      nome:
        maxLength: 120
        minLength: 2
        type: string
      senha:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - nome
    - senha
    type: object
  user.UpdateUserRequestDto:
    properties:
      nome:
        maxLength: 120
        minLength: 2
        type: string
      status:
        enum:
        - ativo
        - inativo
        type: string
    type: object
  user.UserResponseDto:
    properties:
      created_at:
        type: string
      email:
        type: string
      nome:
        type: string
      organization_uuid:
        type: string
      status:
        type: string
      updated_at:
        type: string
      uuid:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/domain/identidade/users:
    get:
      description: Lista paginada dos usuários da organization; com workspace_uuid
        lista só quem tem atribuição nele (via atribuição — a tabela de usuário não
        tem workspace)
      parameters:
      - description: UUID do workspace (fallback quando o host não tem subdomínio)
        in: header
        name: X-Workspace-Id
        type: string
      - description: Página (default 1)
        in: query
        name: page
        type: integer
      - description: Itens por página (teto 100)
        in: query
        name: pageSize
        type: integer
      - description: Filtro por nome
        in: query
        name: nome
        type: string
      - description: Filtro por e-mail
        in: query
        name: email
        type: string
      - description: Filtro por status (ativo|inativo)
        in: query
        name: status
        type: string
      - description: Lista só usuários atribuídos a este workspace
        in: query
        name: workspace_uuid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pagination.Response-user_UserResponseDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Lista usuários
      tags:
      - Identidade · Usuário
    post:
      consumes:
      - application/json
      description: Cria usuário na organization autenticada; a senha nunca volta em
        resposta
      parameters:
      - description: UUID do workspace (fallback quando o host não tem subdomínio)
        in: header
        name: X-Workspace-Id
        type: string
      - description: Dados do usuário
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserRequestDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.UserResponseDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "409":
          description: E-mail já cadastrado nesta organization
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Cria um usuário
      tags:
      - Identidade · Usuário
  /api/domain/identidade/users/{uuid}:
    delete:
      description: Remoção lógica no escopo da organization; sessões abertas são revogadas
        na hora
      parameters:
      - description: UUID do workspace (fallback quando o host não tem subdomínio)
        in: header
        name: X-Workspace-Id
        type: string
      - description: UUID do usuário
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: UUID malformado
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Remove um usuário
      tags:
      - Identidade · Usuário
    get:
      description: Devolve o usuário no escopo da organization resolvida — usuário
        de organization alheia não se distingue de inexistente
      parameters:
      - description: UUID do workspace (fallback quando o host não tem subdomínio)
        in: header
        name: X-Workspace-Id
        type: string
      - description: UUID do usuário
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponseDto'
        "400":
          description: UUID malformado
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Consulta um usuário
      tags:
      - Identidade · Usuário
    patch:
      consumes:
      - application/json
      description: Atualização parcial; inativar encerra imediatamente as sessões
        abertas do usuário (revogação persistida dos refresh tokens)
      parameters:
      - description: UUID do workspace (fallback quando o host não tem subdomínio)
        in: header
        name: X-Workspace-Id
        type: string
      - description: UUID do usuário
        in: path
        name: uuid
        required: true
        type: string
      - description: Campos a atualizar
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponseDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "422":
          description: Transição de estado proibida
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Atualiza um usuário
      tags:
      - Identidade · Usuário
  /api/domain/identidade/users/{uuid}/atribuicoes:
    get:
      description: Vínculos user × workspace × papel do usuário no escopo da organization,
        com o nome canônico de cada papel
      parameters:
      - description: UUID do workspace (fallback quando o host não tem subdomínio)
        in: header
        name: X-Workspace-Id
        type: string
      - description: UUID do usuário
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.AtribuicaoResponseDto'
            type: array
        "400":
          description: UUID malformado
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Lista as atribuições do usuário
      tags:
      - Identidade · Usuário
    post:
      consumes:
      - application/json
      description: 'Papéis são POR workspace: dar poder a alguém é operação própria,
        separada de editar. O workspace precisa pertencer à organization e estar ativo'
      parameters:
      - description: UUID do workspace (fallback quando o host não tem subdomínio)
        in: header
        name: X-Workspace-Id
        type: string
      - description: UUID do usuário
        in: path
        name: uuid
        required: true
        type: string
      - description: Workspace e papel da atribuição
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.AtribuirPapelRequestDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.AtribuicaoResponseDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "404":
          description: Usuário ou papel não encontrado
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "409":
          description: Atribuição duplicada
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "422":
          description: Workspace inválido para esta organization
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Atribui um papel ao usuário em um workspace
      tags:
      - Identidade · Usuário
  /api/domain/identidade/users/{uuid}/atribuicoes/{atribuicaoUuid}:
    delete:
      description: Retira do usuário o papel exercido no workspace identificado pela
        atribuição
      parameters:
      - description: UUID do workspace (fallback quando o host não tem subdomínio)
        in: header
        name: X-Workspace-Id
        type: string
      - description: UUID do usuário
        in: path
        name: uuid
        required: true
        type: string
      - description: UUID da atribuição
        in: path
        name: atribuicaoUuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: UUID malformado
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Remove uma atribuição do usuário
      tags:
      - Identidade · Usuário
...
swagger: "2.0"